requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
googlesearch-python==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ollama==0.4.7
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
